Finalizando CRUD de books

+++ app/Http/Controllers/API/BookController.php
@@ -73,11 +73,59 @@ class BookController extends Controller
     public function delete($id){
         $book = Book::find($id);
         // verificar se existe uma reserva associada
-        $book->delete();
-        return response()->json('Livro removido com sucesso');
+        $list_Reservations = $this->listReservations($id);
+        if(empty($list_Reservations)){
+            try {
+                //code...
+                //$book->reservationsBook()->delete();
+                //$user = $book->user();
+                //$user->reservations()->delete();
+                $book->delete();
+                //$reservation = Reservation::find($id);
+                ///$reservation->delete();
+                $response = [
+                    'success' => true,
+                    'message' => 'Livro removido com sucesso',
+
+                ];
+                return response()->json($response, 200);
+            } catch (\Throwable $th) {
+                //throw $th;
+                return response()->json($th->getMessage(), 500);
+            }
+        }else{
+            $response = [
+                'success' => false,
+                'message' => 'Este livro ainda possui reservas',
+
+            ];
+            return response()->json($response, 200);
+        }
+
 
 
     }
+
+    private function listReservations($id){
+
+        $dateToday = date_create();
+        $book = Book::find($id);
+
+        $list_reservation = array();
+
+        foreach($book->reservationsBook as $reservation){
+
+            $date_devolution = date_create($reservation->pivot->date_devolution);
+            if($dateToday < $date_devolution){
+
+                $item= ['name' => $reservation->name, 'date_reservation' => $reservation->pivot->date_reservation, 'date_devolution' => $reservation->pivot->date_devolution];
+                array_push($list_reservation, $item);
+            }
+        }
+
+        return $list_reservation;
+    }
+
     public function reservations($id){
 
         $dateToday = date_create();
@@ -90,7 +138,7 @@ class BookController extends Controller
             $date_devolution = date_create($reservation->pivot->date_devolution);
             if($dateToday < $date_devolution){
 
-                $item= array('name' => $reservation->name, 'date_reservation' => $reservation->pivot->date_reservation, 'date_devolution' => $reservation->pivot->date_devolution);
+                $item= ['name' => $reservation->name, 'date_reservation' => $reservation->pivot->date_reservation, 'date_devolution' => $reservation->pivot->date_devolution];
                 array_push($list_reservation, $item);
             }
         }
